requirements.txt
opencv-python>=4.8.0
google-genai>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
